requirements.txt
alembic==1.18.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
cryptography==44.0.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
fastapi==0.123.10
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-adk==1.23.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.134.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.0
google-cloud-pubsub==2.34.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.0
google-cloud-storage==3.8.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.60.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.15.0
nassl==5.4.0
neo4j==6.1.0
networkx>=3.0,<3.5
numpy>=1.24.0,<2.3
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api
opentelemetry-exporter-gcp-logging
opentelemetry-exporter-gcp-monitoring
opentelemetry-exporter-gcp-trace
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-resourcedetector-gcp
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson==3.11.5
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
posthog==5.4.0
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
psycopg2-binary==2.9.10
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
reportlab==4.2.5
rich==14.3.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
sslyze==6.3.0
starlette==0.50.0

structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
tls_parser==2.0.2
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
fastapi
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0

matplotlib>=3.6
python-nmap>=0.7.1
